refactor: 由供应商页面持有模型草稿和探测生命周期
